requirements.txt
fastapi
uvicorn
psutil
gputil
setuptools
requests
python-dotenv